Moved all commands into a single pegnet entrypoint
