+++ pegnet/main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/pegnet/pegnet/pncli/cmd"
+import "github.com/pegnet/pegnet/pegnet/cmd"
 
 func main() {
 	cmd.Execute()
